requirements.txt
fastapi
uvicorn
jinja2
weasyprint
pyephem
python-dotenv
mangum